Create audit table for core demography
