+++ src/main/java/uk/ac/ucl/rits/inform/informdb/demographics/AuditCoreDemographic.java
@@ -6,24 +6,22 @@ import uk.ac.ucl.rits.inform.informdb.AuditCore;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
-import javax.persistence.Inheritance;
-import javax.persistence.InheritanceType;
+import javax.persistence.Table;
 import java.time.Instant;
 
 @Entity
-@Inheritance(strategy = InheritanceType.TABLE_PER_CLASS)
+@Table
 @Data
 @EqualsAndHashCode(callSuper = true)
-public class AuditCoreDemographic extends CoreDemographic implements AuditCore<CoreDemographic> {
+public class AuditCoreDemographic extends CoreDemographicParent implements AuditCore<CoreDemographicParent> {
     private static final long serialVersionUID = -8516988957488992519L;
-    @Column(nullable = false, columnDefinition = "timestamp with time zone")
-    private Instant validFrom;
+    @Column
+    private long originalEntityId;
     @Column(columnDefinition = "timestamp with time zone")
     private Instant validUntil;
-    @Column(nullable = false, columnDefinition = "timestamp with time zone")
-    private Instant storedFrom;
     @Column(columnDefinition = "timestamp with time zone")
     private Instant storedUntil;
+
 
     /**
      * Default constructor.
@@ -32,15 +30,16 @@ public class AuditCoreDemographic extends CoreDemographic implements AuditCore<C
     }
 
     /**
-     * Constructor from original entity to be audit logged.
-     * @param originalEntity original entity.
+     * Constructor from original entity and invalidation times.
+     * @param originalEntity original entity to be audited.
+     * @param storedUntil    the time that this change is being made in the DB
+     * @param validUntil     the time at which this fact stopped being true,
+     *                       can be any amount of time in the past
      */
-    public AuditCoreDemographic(CoreDemographic originalEntity) {
+    public AuditCoreDemographic(final CoreDemographic originalEntity, final Instant validUntil, final Instant storedUntil) {
         super(originalEntity);
-    }
-
-    @Override
-    public AuditCore<CoreDemographic> buildFromOriginalEntity(CoreDemographic originalEntity) {
-        return new AuditCoreDemographic(originalEntity);
+        this.validUntil = validUntil;
+        this.storedUntil = storedUntil;
+        this.originalEntityId = originalEntity.getCoreDemographicId();
     }
 }
